% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtils.R
\name{areSimilar}
\alias{areSimilar}
\title{Determines if two strings are similar}
\usage{
areSimilar(string1, string2, threshold)
}
\arguments{
\item{string1}{first string}

\item{string2}{second string}

\item{threshold}{Threshold for similarity in terms of the normalized Levenshtein distance}
}
\value{
Returns TRUE, if the distance is lower or equal the threshold
}
\description{
Determines if two strings are similar
}
