% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{createAlignmentMatrix}
\alias{createAlignmentMatrix}
\title{Creates an alignment matrix for multiple meaning-classes}
\usage{
createAlignmentMatrix(languages, words = 1:210, ...)
}
\arguments{
\item{languages}{numeric vector of language-IDs}

\item{words}{numeric vector of meaning-class-ids}

\item{...}{further arguments passed to createWordAlignment}
}
\value{
returns a concatenated matrix of word-alignments
}
\description{
Creates an alignment matrix for multiple meaning-classes
}
\seealso{
\code{\link{createWordAlignment}}
}
