% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nexus.R
\name{createNexusFile}
\alias{createNexusFile}
\title{Create Nexus file for phylogenetic analyses}
\usage{
createNexusFile(languages, file, words = 1:210, sortBins = FALSE, ...)
}
\arguments{
\item{languages}{Language ids that should be included}

\item{file}{output-file}

\item{words}{word-id that should be included as numeric vector (default is 1:210, which is all words)}

\item{sortBins}{if TRUE, charsets are sorted by their size}

\item{...}{further arguments passed to createAlignmentMatrix}
}
\description{
Create Nexus file for phylogenetic analyses
}
