% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{createWordAlignment}
\alias{createWordAlignment}
\title{Create an alignment-matrix of a meaning-class}
\usage{
createWordAlignment(languages, wordID, ascertainment = TRUE, ...)
}
\arguments{
\item{languages}{The language-IDs as a numeric vector}

\item{wordID}{the ID of the meaning-class (must be a single numeric, cannot process a vector of IDs)}

\item{ascertainment}{if TRUE, an ascertainment column is created at the beginning of the alignment}

\item{...}{further arguments passed to parseCognates}
}
\value{
Returns a binary matrix. First column is ascertainment (if ascertainment is set to TRUE), further columns are cognate-columns and the last columns are singletons
}
\description{
Create an alignment-matrix of a meaning-class
}
\seealso{
\code{\link{createAlignmentMatrix}}
}
