% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.R
\name{download_abvd}
\alias{download_abvd}
\title{Download from ABVD-Database}
\usage{
download_abvd(id, folder = NULL)
}
\arguments{
\item{id}{id of the language that is to be downloaded}

\item{folder}{destination folder of the download. Optional, if default folder in options(abvdLanguageFolder) is specified.}
}
\description{
Download from ABVD-Database
}
