% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastFalseMismatches.R
\name{getAllFalseMismatches}
\alias{getAllFalseMismatches}
\title{False cognate assignments of all languages}
\usage{
getAllFalseMismatches(language_ids, threshold = 0.34, silent = TRUE,
  detailed = FALSE, ...)
}
\arguments{
\item{language_ids}{IDs of all languages in comparison}

\item{threshold}{Threshold, when lexemes are classified as similar. Normalized levensthein-distance, see \code{\link{normalizedLevensthein}}.}

\item{silent}{If progress should be pasted}

\item{detailed}{boolean. If true, additional information is included in the output (takes a bit longer)}

\item{...}{further arguments passed to normalizedLevenshtein}
}
\value{
Returns a data.frame of all lexemes that are suspicious
}
\description{
Detects all pairs of lexemes that are similar, but have different cognate coding. This may take a while for a lot of languages
}
