% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getAuthor}
\alias{getAuthor}
\title{Get the author of a doculect}
\usage{
getAuthor(id, ...)
}
\arguments{
\item{id}{id of the language}

\item{...}{further arguments passed to getWordFile, e.g. if language should be downloaded}
}
\value{
Returns the author as string.
}
\description{
Get the author of a doculect
}
