% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getClade}
\alias{getClade}
\title{Get the classification clade of a language}
\usage{
getClade(id, ...)
}
\arguments{
\item{id}{id of the language}

\item{...}{further arguments passed to getWordFile, e.g. if language should be downloaded}
}
\value{
Returns the classification of the language as string.
}
\description{
Get the classification clade of a language
}
