% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concepts.R
\name{getConcept}
\alias{getConcept}
\title{Get Concept}
\usage{
getConcept(id)
}
\arguments{
\item{id}{Id of the concept in the abvd-database}
}
\value{
returns a string with the name of the concept
}
\description{
Get Concept
}
