% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getGlottocode}
\alias{getGlottocode}
\title{Get the glottocode of a doculect}
\usage{
getGlottocode(id, ...)
}
\arguments{
\item{id}{id of the language}

\item{...}{further arguments passed to getWordFile, e.g. if language should be downloaded}
}
\value{
Returns the glottocode as string.
}
\description{
Get the glottocode of a doculect
}
