% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLanguage.R
\name{getLanguageFile}
\alias{getLanguageFile}
\title{Get csv-file of a language}
\usage{
getLanguageFile(id, download = TRUE)
}
\arguments{
\item{id}{Id of the language}

\item{download}{if TRUE, automatically download the language files, if they cannot be found}
}
\value{
Returns the path of a csv data file.
}
\description{
Get csv-file of a language
}
