% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalOptions.R
\name{getLanguageFolder}
\alias{getLanguageFolder}
\title{Get Language folder}
\usage{
getLanguageFolder()
}
\value{
Returns the folder, where language files are stores. Should be specified with options(abvdLanguageFolder = "~/path/to/my/languages")
}
\description{
Get Language folder
}
