% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLanguage.R
\name{getLanguageName}
\alias{getLanguageName}
\title{Get language name}
\usage{
getLanguageName(id, ...)
}
\arguments{
\item{id}{id of the language}

\item{...}{further arguments passed to getWordFile, e.g. if language should be downloaded}
}
\value{
Returns name of the language. Empty spaces and commas are stripped.
}
\description{
Get language name
}
