% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLanguage.R
\name{getWordFile}
\alias{getWordFile}
\title{Get wordlist file of a language}
\usage{
getWordFile(id, download = TRUE)
}
\arguments{
\item{id}{Id of the language}

\item{download}{if TRUE, automatically download the language files, if they cannot be found}
}
\value{
Returns the path of a wordlist.
}
\description{
Get wordlist file of a language
}
