% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalOptions.R
\name{getWordList}
\alias{getWordList}
\title{Get Word list}
\usage{
getWordList()
}
\value{
returns path to wordlist.csv
}
\description{
word list must be manually downloaded once and then specified with options(abvdWordList="~/path/to/wordlist.csv")
}
