% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtils.R
\name{normalizedLevenshtein}
\alias{normalizedLevenshtein}
\title{Normalized Levenshtein distance}
\usage{
normalizedLevenshtein(string1, string2)
}
\arguments{
\item{string1}{First string of comparison}

\item{string2}{Second string of comparison}
}
\value{
Returns numeric value of the normalized Levenshtein distance
}
\description{
Levenshtein distance of two strings is the minimal amount of edits (insertions, deletions and substitutions) to go from string1 to string2. This is normalized by the length of the longer string.
}
