% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedCognateDistance.R
\name{pairwiseSCDistance}
\alias{pairwiseSCDistance}
\title{Pairwise shared Cognate distance}
\usage{
pairwiseSCDistance(lang1, lang2, alignment)
}
\arguments{
\item{lang1}{row index of the first language in the alignment}

\item{lang2}{row index of the second language in the alignment}

\item{alignment}{alignment matrix as created by createAlignmentMatrix()}
}
\value{

}
\description{
This calculates a distance based on the amount of shared cognates. Two languages share a cognate for a word
if they have at least one cognate code in common for that word (e.g. if language 1 has for the concept cognate "1, 2" and language 2
has "1, 3", they share a cognate). The distance is then the proportion of concepts, where these two languages do not share a cognate.
}
