% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLanguage.R
\name{readLanguage}
\alias{readLanguage}
\alias{readMultipleLanguages}
\title{Read Language}
\usage{
readLanguage(id, download = TRUE)

readMultipleLanguages(ids, download = TRUE)
}
\arguments{
\item{id}{Id of the language}

\item{download}{if TRUE, automatically download the language files, if they cannot be found}

\item{ids}{numerical vector of language-ids}
}
\value{
Returns a data-frame of the wordlist

Returns a list of language-dataframes.
}
\description{
Read Language

Reads multiple languages
}
\section{Functions}{
\itemize{
\item \code{readMultipleLanguages}: Reads multiple languages and outputs them as a list
}}

