% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareLanguages.R
\name{sumFalseMismatches}
\alias{sumFalseMismatches}
\title{Amount of false mismatches of given languages (Deprecated)}
\usage{
sumFalseMismatches(languages, threshold = 0.4)
}
\arguments{
\item{languages}{Languages to compare as numerical vector of ids}

\item{threshold}{Threshold, when lexemes are classified as similar. Normalized levensthein-distance, see \code{\link{normalizedLevenshtein}}.}
}
\description{
This is too slow, dont use! Use \code{sum\link{getAllFalseMismatches})}
}
