% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareLanguages.R
\name{twoLanguageWordlist}
\alias{twoLanguageWordlist}
\title{Create a comparable word list of two languages}
\usage{
twoLanguageWordlist(lang1, lang2, concepts = NULL, ...)
}
\arguments{
\item{lang1}{First language (ID or data-frame generated from a csv-fiel via \code{\link{read.csv}})}

\item{lang2}{Second language}

\item{concepts}{Optional input. Character-vector of all 210 concepts. Can additionally be provided to speed things up}
}
\value{
Returns a dataframe with all lexemes sorted by meaning-class-ID with their corresponding cognate classes.
}
\description{
Create a comparable word list of two languages
}
